backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
redis[hiredis]==5.0.1
motor==3.3.2
pymongo==4.6.0
pydantic
pydantic-settings
pydantic[email]
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
qrcode[pil]==7.4.2
Pillow
pymongo[srv]
certifi
requests
